Add max retries, macros and docs
